requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
crispy-bootstrap4==2023.1
crispy-bootstrap5==0.7
Django==4.2.5
django-bootstrap4==23.2
django-crispy-forms==2.1
django-debug-toolbar==4.2.0
djangorestframework==3.14.0
Pillow==10.1.0
pytz==2024.1
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
